+++ server-php/api/Models/Itinerary.php
@@ -1,11 +1,8 @@
 <?php
 
-namespace ItineraryApi\Models;
+namespace Api\Models;
 
 use Illuminate\Database\Eloquent\Model as Model;
-use ItineraryApi\Utils\UnauthorizedException;
-
-use function PHPSTORM_META\type;
 
 class Itinerary extends Model
 {
@@ -15,14 +12,14 @@ class Itinerary extends Model
 
     protected $hidden = ['user_id', 'created_at', 'updated_at'];
 
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'id');
+    }
+
     public function activities()
     {
         return $this->hasMany(Activity::class, 'activity_id');
-    }
-
-    public function user()
-    {
-        return $this->belongsTo(User::class, 'user_id', 'id');
     }
 
     public static function getItineraries($userId)
@@ -30,32 +27,32 @@ class Itinerary extends Model
         return self::where('user_id', $userId)->get();
     }
 
-    public static function getItinerary($id, $userId)
+    public static function getItineraryById($id, $userId)
     {
-        $itinerary = self::findOrFail($id);
-        UnauthorizedException::checkId($itinerary->user_id, $userId);
-
-        return $itinerary;
+        return self::where('id', $id)->where('user_id', $userId)->first();
     }
 
-    public static function createItinerary($userId)
+    public static function createItinerary($userId, $body)
     {
         $itinerary = new Itinerary();
         $itinerary->user_id = $userId;
-        $itinerary->title = $_REQUEST['title'];
-        $itinerary->start_date = $_REQUEST['start_date'];
-        $itinerary->end_date = $_REQUEST['end_date'];
+        $itinerary->title = $body['title'];
+        $itinerary->start_date = $body['start_date'];
+        $itinerary->end_date = $body['end_date'];
         $itinerary->save();
 
         return $itinerary;
     }
 
-    public static function updateItinerary($id, $userId)
+    public static function updateItinerary($id, $userId, $body)
     {
-        $itinerary = self::findOrFail($id);
-        UnauthorizedException::checkId($itinerary->user_id, $userId);
+        $itinerary = self::getItineraryById($id, $userId);
 
-        foreach ($_REQUEST as $field => $value) {
+        if (!$itinerary) {
+            return false;
+        }
+
+        foreach ($body as $field => $value) {
             $itinerary->$field = $value;
         }
 
@@ -66,9 +63,13 @@ class Itinerary extends Model
 
     public static function deleteItinerary($id, $userId)
     {
-        $itinerary = self::findOrFail($id);
-        UnauthorizedException::checkId($itinerary->user_id, $userId);
+        $itinerary = self::getItineraryById($id, $userId);
+
+        if (!$itinerary) {
+            return false;
+        }
 
         $itinerary->delete();
+        return true;
     }
 }
